+++ tests/cli/add/test_add_variants.py
@@ -9,14 +9,19 @@ def test_add_variants_no_dataset(mock_app):
     runner = mock_app.test_cli_runner()
 
     # When invoking the add variants with a random dataset ID
-    result =  runner.invoke(cli, [
-        'add',
-        'variants',
-        '-ds', 'a_dataset',
-        '-vcf', test_snv_vcf_path,
-        '-sample', 'a_sample',
-        '-type', 'snv'
-    ])
+    result = runner.invoke(
+        cli,
+        [
+            "add",
+            "variants",
+            "-ds",
+            "a_dataset",
+            "-vcf",
+            test_snv_vcf_path,
+            "-sample",
+            "a_sample",
+        ],
+    )
 
     # Then the command should return error
     assert result.exit_code == 1
@@ -34,19 +39,24 @@ def test_add_variants_empty_vcf(mock_app, test_dataset_cli, database):
     database["dataset"].insert_one(dataset)
 
     # When invoking the add variants with an existing but empty VCF file
-    result =  runner.invoke(cli, [
-        'add',
-        'variants',
-        '-ds', dataset["_id"],
-        '-vcf', test_empty_vcf_path,
-        '-sample', 'ADM1059A2',
-        '-type', 'sv'
-    ])
+    result = runner.invoke(
+        cli,
+        [
+            "add",
+            "variants",
+            "-ds",
+            dataset["_id"],
+            "-vcf",
+            test_empty_vcf_path,
+            "-sample",
+            "ADM1059A2",
+        ],
+    )
 
     # Then the command should return error
     assert result.exit_code == 1
     # And a specific error message
-    assert f"Coundn't extract any variant from the provided file" in result.output
+    assert f"Provided VCF file doesn't contain any variant" in result.output
 
 
 def test_add_variants_wrong_samples(mock_app, test_dataset_cli, database):
@@ -59,18 +69,26 @@ def test_add_variants_wrong_samples(mock_app, test_dataset_cli, database):
     database["dataset"].insert_one(dataset)
 
     # When invoking the add variants for a sample not in the VCF file
-    result =  runner.invoke(cli, [
-        'add',
-        'variants',
-        '-ds', dataset["_id"],
-        '-vcf', test_snv_vcf_path,
-        '-sample', 'a_sample',
-        '-type', 'snv'
-    ])
+    result = runner.invoke(
+        cli,
+        [
+            "add",
+            "variants",
+            "-ds",
+            dataset["_id"],
+            "-vcf",
+            test_snv_vcf_path,
+            "-sample",
+            "a_sample",
+        ],
+    )
     # Then the command should return error
     assert result.exit_code == 1
     # And a specific error message
-    assert f"Error. One or more provided samples are not contained in the VCF file" in result.output
+    assert (
+        f"Error. One or more provided samples are not contained in the VCF file"
+        in result.output
+    )
 
 
 def test_add_variants_snv_vcf(mock_app, test_dataset_cli, database):
@@ -81,20 +99,37 @@ def test_add_variants_snv_vcf(mock_app, test_dataset_cli, database):
     # Having a database containing a dataset
     dataset = test_dataset_cli
     database["dataset"].insert_one(dataset)
+    sample = "ADM1059A1"
 
     # When invoking the add variants from a VCF file
-    result =  runner.invoke(cli, [
-        'add',
-        'variants',
-        '-ds', dataset["_id"],
-        '-vcf', test_snv_vcf_path,
-        '-sample', 'ADM1059A1',
-        '-type', 'snv'
-    ])
+    result = runner.invoke(
+        cli,
+        [
+            "add",
+            "variants",
+            "-ds",
+            dataset["_id"],
+            "-vcf",
+            test_snv_vcf_path,
+            "-sample",
+            sample,
+        ],
+    )
 
     # Then the command should NOT return error
     assert result.exit_code == 0
     assert f"variants loaded into the database" in result.output
 
-
-    #
+    # and variants parsed correctly have been saved to database
+    test_variant = database["variant"].find_one()
+    assert isinstance(test_variant["referenceName"], str)
+    assert isinstance(test_variant["start"], int)
+    assert isinstance(test_variant["startMin"], int)
+    assert isinstance(test_variant["startMax"], int)
+    assert isinstance(test_variant["end"], int)
+    assert isinstance(test_variant["endMin"], int)
+    assert isinstance(test_variant["endMax"], int)
+    assert isinstance(test_variant["referenceBases"], str)
+    assert isinstance(test_variant["alternateBases"], str)
+    assert test_variant["assemblyId"] == "GRCh37"
+    assert test_variant["datasetIds"] == [{dataset["_id"]: {"samples": [sample]}}]