+++ tests/conftest.py
@@ -5,21 +5,26 @@ from cgbeacon2.server import create_app
 
 DATABASE_NAME = "testdb"
 
-@pytest.fixture(scope='function')
+
+@pytest.fixture(scope="function")
 def pymongo_client(request):
     """Get a client to the mongo database"""
     mock_client = mongomock.MongoClient()
+
     def teardown():
         mock_client.drop_database(DATABASE_NAME)
+
     request.addfinalizer(teardown)
     return mock_client
 
-@pytest.fixture(scope='function')
+
+@pytest.fixture(scope="function")
 def database(request, pymongo_client):
     """Get an adapter connected to mongo database"""
     mongo_client = pymongo_client
     database = mongo_client[DATABASE_NAME]
     return database
+
 
 @pytest.fixture
 def mock_app(database):
@@ -28,16 +33,17 @@ def mock_app(database):
     app.db = database
     return app
 
+
 @pytest.fixture
 def test_dataset_cli():
     """A test dataset dictionary"""
     dataset = dict(
-        _id = "test-dataset",
-        name = "Test dataset",
-        assembly_id = "GRCh37",
-        description = "Test dataset description",
-        version = 1.0,
-        url = "external_url.url",
-        consent_code = "HMB"
+        _id="test-dataset",
+        name="Test dataset",
+        assembly_id="GRCh37",
+        description="Test dataset description",
+        version=1.0,
+        url="external_url.url",
+        consent_code="HMB",
     )
     return dataset